refactor(app-core-settings): split lib.rs into keys, validation, read, writes modules
Pure move: the 2327-line crate root keeps its doc header, imports, the
five live mod declarations, and the public re-export surface; key
predicates, value validation, get_settings, and the write/override/
snapshot use cases move verbatim into keys.rs, validation.rs, read.rs,
writes.rs, with the test module in tests.rs.
